Search conferences done. Did a little more code refactor
